Created 2 Directories, added Python impl
The C++ project is largely the same, the only difference is that I added the Python script with support for negative Multibrot sets. I will likely use that as basis for my C++ solution.

The python solution can also plot a normal distribution of how many iterations are required on average in order to get a clear view of how many times one should test for periodicity before giving up.